feat: Add parse func tests
